api/requirements.txt
fastapi
python-multipart
uvicorn
networkx
osmnx
geopandas
scikit-learn
opencv-python
torch
transformers
Pillow
plyfile
pytest
httpx
pytest-asyncio
